fix int

+++ db/migrate/20120501040221_create_apps.rb
@@ -2,7 +2,7 @@ class CreateApps < ActiveRecord::Migration
   def self.up
     create_table :apps do |t|
       t.integer :user_id
-      t.integer :app_id
+      t.double :app_id
       t.string :app_secret
       t.string :expiration_date
       t.string :app_name
